requirements.txt
boto3
ansible_vault
ruamel.yaml
ansible
pylint